64-bit work in progress
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@997 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCMView.m
@@ -7207,10 +7207,10 @@ BOOL	lowRes = NO;
 	[self setNeedsDisplay:YES];
 }
 
--(void)setTag:(int)aTag{
+-(void)setTag:(NSInteger)aTag{
 	_tag = aTag;
 }
-- (int)tag{
+- (NSInteger)tag{
 	return _tag;
 }
 -(float)curWW{
